+++ backend/.env
@@ -1,0 +1,5 @@
+PORT=80
+
+BASE_URL=http://localhost
+
+SPOONACULAR_API_KEY=34b7bc821685421198bfdd8805208f43
